requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.1.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
exceptiongroup==1.2.2
fake-useragent==2.0.3
frozenlist==1.5.0
google-auth==2.38.0
google-genai==1.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
multidict==6.1.0
openai==1.61.1
playwright==1.49.1
propcache==0.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.0.0
PySocks==1.7.1
pytz==2024.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
stem==1.8.2
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
websockets==14.2
yarl==1.18.3
